Discussion 2 Practice
As part of our discussion for this week, I looked through additional zip files and R projects to see if I could reproduce the analysis. I was successful with a couple and unsuccessful with several. I deleted several of the zip files that I was unable to reproduce to avoid clutter.